refactor(frontend): move logout button to right location
